動く状態

+++ python/provision/python.sh
@@ -4,9 +4,9 @@ set -Ceu
 # Python公式
 # https://www.python.jp/install/ubuntu/index.html
 
-sudo apt-get update
+sudo apt update
 
-sudo apt-get install -y build-essential \
+sudo apt install -y build-essential \
                         libncursesw5-dev \
                         libgdbm-dev \
                         libc6-dev \
@@ -23,19 +23,23 @@ sudo apt-get install -y build-essential \
 USER="ubuntu"
 HOME_DIRECTORY="/home/${USER}"
 PYENV_HOME="/home/${USER}/.pyenv"
-PYENV_SHELL_PATH=${HOME_DIRECTORY}"/.profile"
+PYENV_SHELL_PATH=${HOME_DIRECTORY}"/.bash_profile"
 PYENV_COMMAND="source "${PYENV_SHELL_PATH}
 
-sudo git clone https://github.com/yyuu/pyenv.git ${PYENV_HOME}
-sudo echo 'export PYENV_ROOT="$HOME/.pyenv"' >> ${HOME_DIRECTORY}/.profile
-sudo echo 'export PATH="$PYENV_ROOT/bin:$PATH"' >> ${HOME_DIRECTORY}/.profile
-sudo echo 'eval "$(pyenv init -)"' >> ${HOME_DIRECTORY}/.profile
+sudo chmod 0664 ${PYENV_SHELL_PATH}
+sudo chown ubuntu:ubuntu ${PYENV_SHELL_PATH}
 
-sudo su -l ${USER} -c "${PYENV_COMMAND}"
-sudo su -l ${USER} -c "pyenv install 3.6.0"
-sudo su -l ${USER} -c "pyenv global 3.6.0"
+if [ ! -e ${PYENV_HOME} ]; then
+  sudo su -l ${USER} -c "git clone https://github.com/pyenv/pyenv.git ${PYENV_HOME}"
+
+  sudo su -l ${USER} -c "${PYENV_COMMAND}"
+  sudo su -l ${USER} -c "pyenv install 3.6.0"
+  sudo su -l ${USER} -c "pyenv global 3.6.0"
+
+  sudo su -l ${USER} -c "pip install requests"
+  sudo su -l ${USER} -c "pip install lxml"
+  sudo su -l ${USER} -c "pip install cssselect"
+fi
 
 
-sudo su -l ${USER} -c "pip install requests"
-sudo su -l ${USER} -c "pip install lxml"
-sudo su -l ${USER} -c "pip install cssselect"
+
